created a GameInviteModal; it is called now in chat; game invites through chat works
