+++ front-vite/src/Pages/Channels/index.tsx
@@ -15,10 +15,14 @@ import { Add as AddIcon, Group as GroupIcon, Cancel as CancelIcon, Logout as Log
 import { timeStamp } from 'console';
 import { index } from 'cheerio/dist/commonjs/api/traversing';
 import { useChatContext, socket } from '../../Layout/Chat/ChatContext';
-import { User, useUser } from '../../Providers/UserContext/User';
+import { useUser } from '../../Providers/UserContext/User';
+import { User } from '../../Types/User/Interfaces';
 import { getAll } from '../../Providers/UserContext/User';
 import { getRandomValues } from 'crypto';
 import { copyFileSync } from 'fs';
+import { GameInviteModal } from '../Game/inviteModal';
+import { PowerUpSelected } from '../../Types/Game/Enum';
+import { inviteToGame } from '../../Providers/NotificationContext/Notification';
 // import { Socket } from 'socket.io-client';
 
 
@@ -73,7 +77,9 @@ const ChannelsPage: React.FC = () => {
 	const [joinedChannels, setJoinedChannels] = useState<ChatRoom[]>([]);
 	const [selectedChannel, setSelectedChannel] = useState<ChatRoom | null>(null);
 	const [selectedAvailableChannel, setSelectedAvailableChannel] = useState<ChatRoom | null>(null);
-  
+	const [powerupValue, setPowerupValue] = useState<PowerUpSelected>(0);
+	const [modalOpen, setModalOpen] = useState<Boolean>(false);
+
 	useEffect(() => {
 		// if (chatProps.chatRooms) {
 		const joined = chatProps.chatRooms.filter((channel) =>
@@ -758,10 +764,35 @@ const ChannelsPage: React.FC = () => {
 	};
 //////////////////////////////////////////////////////////////////////
 
+	const handleModalClose = () => {
+		setModalOpen(false);
+	};
+
+	const handleModalOpen = () => {
+		setModalOpen(true); 
+	};
+
+	const checkIfBlocked = (otherUser: User) =>
+	{
+		if (user.blocked.find((blockedId:string) => blockedId === otherUser.intraId.toString())) 
+		{
+			return (true);
+		}
+		return (false);
+	}
+
+	const InviteGame = (otherUser: User) => {
+		
+		handleModalClose();
+		if (checkIfBlocked(otherUser) == true)
+				return ;
+
+		inviteToGame(user.id.toString(), otherUser.id.toString(), powerupValue);
+	}
+
 	const UserLine: React.FC<{eveuser: User}> = ({user}) => {
 		return (
 			<Stack
-				onClick={() => {(navigate(`/profile/1`))}} // TO BE REPLACED!
 				direction={'row'}
 				paddingX={'0.5em'}
 				justifyContent={'center'}
@@ -783,16 +814,25 @@ const ChannelsPage: React.FC = () => {
 				}}
 			>
 				<AccountCircleIcon sx={{ marginRight: 1}}/>
-				<Typography noWrap sx={{ maxWidth: '78%', overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
+				<Typography noWrap onClick={() => {(navigate(`/profile/` + user.id.toString()))}} sx={{ maxWidth: '78%', overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap' }}>
 					{user.nameIntra}
 		 		 </Typography>
-				<Box sx={{ flexGrow: 1 }} />
-				<IconButton
-					onClick={handleSendGameInvite}
-					sx={{  }}
-				>
-					<GameIcon sx={{ }}/>
-				</IconButton>
+				<Box sx={{ flexGrow: 1 }} /> 
+				<Tooltip title='Send game invite' arrow>
+					<IconButton
+						onClick={handleModalOpen}
+						sx={{  }}
+					>
+						<GameIcon sx={{ }}/>
+					</IconButton>
+				</Tooltip>
+				{modalOpen && 
+					<GameInviteModal 
+						open={modalOpen} 
+						onClose={() => InviteGame(user)} 
+						setValue={(revalue: PowerUpSelected) => {setPowerupValue(revalue)}} 
+					/>
+				}
 				<Tooltip title='Send a direct messsage' arrow>
 					<IconButton
 						onClick={(event: React.MouseEvent) => handleSendDirectMessageClick(event, user)}
@@ -808,15 +848,12 @@ const ChannelsPage: React.FC = () => {
 
 	const renderUsers = () => (
 		<Stack gap={1}>
-			{users.map((user) => (
-				<UserLine key={user.id} user={user} />	
-			))}
-			
-			{/* {Array.from({ length: 20 }).map((_, index) => (
-				<UserLine key={index} />	
-			))} */}
+			{users
+				.filter((item: User) => item.intraId !== user.intraId) // ✅ Filter users first
+				.map((item: User) => <UserLine key={item.id} user={item} />)}
 		</Stack>
 	);
+
 
 	//---Function to render the list of channels---//
 	const renderJoinedChannels = (channels: ChatRoom[]) => {
